\name{ca65}
\alias{ca65}
\docType{data}
\title{
Sample Case Data
}
\description{
Sample case from the dataset described in Ressom et al. (2007) which consists of MALDI-TOF MS spectra that extend from 800 to 10000 Da obtained from serum samples of 72 patients with HCC and 72 healthy individuals. HCC or Hepatocellular carcinoma (also called malignant hepatoma) is a primary cancer of the liver.
}
\usage{data(ca65)}
\format{
  A data frame with 136680 observations on the following 2 variables.
  \describe{
    \item{\code{m.z}}{mass to charge values}
    \item{\code{y}}{corresponding intensities}
  }
}
\details{
Sample data from patient with HCC.
}
\source{
http://microarray.georgetown.edu/web/files/
}
\references{
Ressom HW, Varghese RS, Drake SK, Hortin GL, Abdel-Hamid M, Loffredo CA and Goldman R (2007) Peak selection from MALTI-TOF mass spectra using ant colony optimization BIOINFORMATICS 23(5):619-626 http://bioinformatics.oxfordjournals.org/cgi/reprint/23/5/619
}
\keyword{datasets}

