\name{pkDACLASS-package}
\alias{pkDACLASS-package}
\alias{pkDACLASS}
\docType{package}
\title{
A complete statistical analysis of MALDI-TOF Mass Spectrometry Proteomics data
}
\description{
The package pkDACLASS provides unique monoisotopic peaks of peptides. It will align the detected monoisotopic peaks from multiple samples and classify the samples with these peaks and report the classification accuracy. pkDACLASS is able to conduct statistical test for differential analysis and report the differentially expressed monoisotopic peptide peaks for multiple samples.  
}
\details{
\tabular{ll}{
Package: \tab pkDACLASS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-05\cr
License: \tab LGPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Juliet Ndukum <jpntsang@gmail.com>, Mourad Atlas <mourad.atlas@gmail.com>, Susmita Datta <susmdatta@gmail.com>

Maintainer: Juliet Ndukum <jpntsang@gmail.com>
}
\references{
Atlas M, Datta S (2009) A Statistical Technique for Monoisotopic Peak Detection in a Mass Spectrum. J Proteomics Bioinform 2: 202-216
}
\keyword{ package }

